Newton 3.08 start using the new dgVector class which combine Simd operation with scalar operation. Scalar operation are slower than before but only when using the newer integer and relational operations. This have to be delet with #ifdef /#end in the code.
for reference: the the stacks of  20 x 20 high pyramid of boxes (2100 bodies) in 3.06 was running at 50 mili secunds per physics update. Now when using new vector class with simd it runs at 35 millisecond per frame
and in scalar mode run at 75.
Later I should be able to bring teh scale more to 50 millisecond as well by ysubn #ifdef / #end in the inner loop of the solver.

with this the is not need for special simd code in engine, we can delete all those functions.